requirements.txt
python-telegram-bot
yfinance
openai
python-dotenv
ib_insync
pysqlite3-binary
textblob
apscheduler
